requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-firebase-authentication==1.0.9
django-heroku==0.3.1
django-likert-field==0.2.0
django-nextjs==2.1.5
djangorestframework==3.14.0
filelock==3.8.0
firebase-admin==6.0.0
frozenlist==1.3.1
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
msgpack==1.0.4
multidict==6.0.2
pbr==5.10.0
platformdirs==2.5.2
proto-plus==1.22.1
protobuf==4.21.7
psycopg2==2.9.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stevedore==4.0.0
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
websockets==10.3
whitenoise==6.2.0
yarl==1.8.1
zope.interface==5.4.0